moved mux drawing functions to seperate file
