+++ vpr/src/draw/draw.cpp
@@ -38,6 +38,8 @@
 #include "draw_xtoy.h"
 #include "draw_toggle_functions.h"
 #include "draw_triangle.h"
+#include "draw_mux.h"
+#include "draw_searchbar.h"
 #include "read_xml_arch_file.h"
 #include "draw_global.h"
 #include "intra_logic_block.h"
@@ -109,15 +111,8 @@ void draw_partial_route(const std::vector<int>& rr_nodes_to_draw,
 
 static int get_track_num(int inode, const vtr::OffsetMatrix<int>& chanx_track, const vtr::OffsetMatrix<int>& chany_track);
 static bool draw_if_net_highlighted(ClusterNetId inet);
-static int draw_check_rr_node_hit(float click_x, float click_y);
-
-static void draw_expand_non_configurable_rr_nodes_recurr(int from_node,
-                                                         std::set<int>& expanded_nodes);
-static bool highlight_rr_nodes(float x, float y);
 static void highlight_blocks(double x, double y);
 
-static inline ezgl::rectangle draw_mux(ezgl::point2d origin, e_side orientation, float height, ezgl::renderer* g);
-static inline ezgl::rectangle draw_mux(ezgl::point2d origin, e_side orientation, float height, float width, float height_scale, ezgl::renderer* g);
 
 static void draw_flyline_timing_edge(ezgl::point2d start, ezgl::point2d end, float incr_delay, ezgl::renderer* g);
 static void draw_routed_timing_edge(tatum::NodeId start_tnode,
@@ -1125,49 +1120,6 @@ void draw_x(float x, float y, float size, ezgl::renderer* g) {
     g->draw_line({x - size, y - size}, {x + size, y + size});
 }
 
-/* This function computes and returns the boundary coordinates of a channel
- * wire segment. This can be used for drawing a wire or determining if a
- * wire has been clicked on by the user.
- * TODO: Fix this for global routing, currently for detailed only.
- */
-ezgl::rectangle draw_get_rr_chan_bbox(int inode) {
-    double left = 0, right = 0, top = 0, bottom = 0;
-    t_draw_coords* draw_coords = get_draw_coords_vars();
-    auto& device_ctx = g_vpr_ctx.device();
-    const auto& rr_graph = device_ctx.rr_graph;
-    auto rr_node = RRNodeId(inode);
-
-    switch (rr_graph.node_type(rr_node)) {
-        case CHANX:
-            left = draw_coords->tile_x[rr_graph.node_xlow(rr_node)];
-            right = draw_coords->tile_x[rr_graph.node_xhigh(rr_node)]
-                    + draw_coords->get_tile_width();
-            bottom = draw_coords->tile_y[rr_graph.node_ylow(rr_node)]
-                     + draw_coords->get_tile_width()
-                     + (1. + rr_graph.node_track_num(rr_node));
-            top = draw_coords->tile_y[rr_graph.node_ylow(rr_node)]
-                  + draw_coords->get_tile_width()
-                  + (1. + rr_graph.node_track_num(rr_node));
-            break;
-        case CHANY:
-            left = draw_coords->tile_x[rr_graph.node_xlow(rr_node)]
-                   + draw_coords->get_tile_width()
-                   + (1. + rr_graph.node_track_num(rr_node));
-            right = draw_coords->tile_x[rr_graph.node_xlow(rr_node)]
-                    + draw_coords->get_tile_width()
-                    + (1. + rr_graph.node_track_num(rr_node));
-            bottom = draw_coords->tile_y[rr_graph.node_ylow(rr_node)];
-            top = draw_coords->tile_y[rr_graph.node_yhigh(rr_node)]
-                  + draw_coords->get_tile_width();
-            break;
-        default:
-            // a problem. leave at default value (ie. zeros)
-            break;
-    }
-    ezgl::rectangle bound_box({left, bottom}, {right, top});
-
-    return bound_box;
-}
 
 /* Returns the coordinates at which the center of this pin should be drawn. *
  * inode gives the node number, and iside gives the side of the clb or pad  *
@@ -1491,206 +1443,6 @@ static bool draw_if_net_highlighted(ClusterNetId inet) {
     return false;
 }
 
-/* If an rr_node has been clicked on, it will be highlighted in MAGENTA.
- * If so, and toggle nets is selected, highlight the whole net in that colour.
- */
-void highlight_nets(char* message, int hit_node) {
-    t_trace* tptr;
-    auto& cluster_ctx = g_vpr_ctx.clustering();
-    auto& route_ctx = g_vpr_ctx.routing();
-
-    t_draw_state* draw_state = get_draw_state_vars();
-
-    for (auto net_id : cluster_ctx.clb_nlist.nets()) {
-        for (tptr = route_ctx.trace[net_id].head; tptr != nullptr;
-             tptr = tptr->next) {
-            if (draw_state->draw_rr_node[tptr->index].color == ezgl::MAGENTA) {
-                draw_state->net_color[net_id] = draw_state->draw_rr_node[tptr->index].color;
-                if (tptr->index == hit_node) {
-                    std::string orig_msg(message);
-                    sprintf(message, "%s  ||  Net: %zu (%s)", orig_msg.c_str(),
-                            size_t(net_id),
-                            cluster_ctx.clb_nlist.net_name(net_id).c_str());
-                }
-            } else if (draw_state->draw_rr_node[tptr->index].color
-                       == ezgl::WHITE) {
-                // If node is de-selected.
-                draw_state->net_color[net_id] = ezgl::BLACK;
-                break;
-            }
-        }
-    }
-    application.update_message(message);
-}
-
-/* If an rr_node has been clicked on, it will be either highlighted in MAGENTA,
- * or de-highlighted in WHITE. If highlighted, and toggle_rr is selected, highlight
- * fan_in into the node in blue and fan_out from the node in red. If de-highlighted,
- * de-highlight its fan_in and fan_out.
- */
-void draw_highlight_fan_in_fan_out(const std::set<int>& nodes) {
-    t_draw_state* draw_state = get_draw_state_vars();
-    auto& device_ctx = g_vpr_ctx.device();
-    const auto& rr_graph = device_ctx.rr_graph;
-    for (auto node : nodes) {
-        /* Highlight the fanout nodes in red. */
-        for (t_edge_size iedge = 0, l = rr_graph.num_edges(RRNodeId(node));
-             iedge < l; iedge++) {
-            int fanout_node = size_t(rr_graph.edge_sink_node(RRNodeId(node), iedge));
-
-            if (draw_state->draw_rr_node[node].color == ezgl::MAGENTA
-                && draw_state->draw_rr_node[fanout_node].color
-                       != ezgl::MAGENTA) {
-                // If node is highlighted, highlight its fanout
-                draw_state->draw_rr_node[fanout_node].color = DRIVES_IT_COLOR;
-                draw_state->draw_rr_node[fanout_node].node_highlighted = true;
-            } else if (draw_state->draw_rr_node[node].color == ezgl::WHITE) {
-                // If node is de-highlighted, de-highlight its fanout
-                draw_state->draw_rr_node[fanout_node].color = DEFAULT_RR_NODE_COLOR;
-                draw_state->draw_rr_node[fanout_node].node_highlighted = false;
-            }
-        }
-
-        /* Highlight the nodes that can fanin to this node in blue. */
-        for (const RRNodeId& inode : rr_graph.nodes()) {
-            for (t_edge_size iedge = 0, l = rr_graph.num_edges(inode); iedge < l;
-                 iedge++) {
-                int fanout_node = size_t(rr_graph.edge_sink_node(inode, iedge));
-                if (fanout_node == node) {
-                    if (draw_state->draw_rr_node[node].color == ezgl::MAGENTA
-                        && draw_state->draw_rr_node[size_t(inode)].color
-                               != ezgl::MAGENTA) {
-                        // If node is highlighted, highlight its fanin
-                        draw_state->draw_rr_node[size_t(inode)].color = ezgl::BLUE;
-                        draw_state->draw_rr_node[size_t(inode)].node_highlighted = true;
-                    } else if (draw_state->draw_rr_node[node].color
-                               == ezgl::WHITE) {
-                        // If node is de-highlighted, de-highlight its fanin
-                        draw_state->draw_rr_node[size_t(inode)].color = DEFAULT_RR_NODE_COLOR;
-                        draw_state->draw_rr_node[size_t(inode)].node_highlighted = false;
-                    }
-                }
-            }
-        }
-    }
-}
-
-/* This is a helper function for highlight_rr_nodes(). It determines whether
- * a routing resource has been clicked on by computing a bounding box for that
- *  and checking if the mouse click hit inside its bounding box.
- *
- *  It returns the hit RR node's ID (or OPEN if no hit)
- */
-static int draw_check_rr_node_hit(float click_x, float click_y) {
-    int hit_node = OPEN;
-    ezgl::rectangle bound_box;
-
-    t_draw_coords* draw_coords = get_draw_coords_vars();
-    auto& device_ctx = g_vpr_ctx.device();
-    const auto& rr_graph = device_ctx.rr_graph;
-
-    for (const RRNodeId& rr_id : device_ctx.rr_graph.nodes()) {
-        size_t inode = (size_t)rr_id;
-        switch (rr_graph.node_type(rr_id)) {
-            case IPIN:
-            case OPIN: {
-                int i = rr_graph.node_xlow(rr_id);
-                int j = rr_graph.node_ylow(rr_id);
-                t_physical_tile_type_ptr type = device_ctx.grid[i][j].type;
-                int width_offset = device_ctx.grid[i][j].width_offset;
-                int height_offset = device_ctx.grid[i][j].height_offset;
-                int ipin = rr_graph.node_pin_num(rr_id);
-                float xcen, ycen;
-                for (const e_side& iside : SIDES) {
-                    // If pin exists on this side of the block, then get pin coordinates
-                    if (type->pinloc[width_offset][height_offset][size_t(iside)][ipin]) {
-                        draw_get_rr_pin_coords(inode, &xcen, &ycen, iside);
-
-                        // Now check if we clicked on this pin
-                        if (click_x >= xcen - draw_coords->pin_size && click_x <= xcen + draw_coords->pin_size && click_y >= ycen - draw_coords->pin_size && click_y <= ycen + draw_coords->pin_size) {
-                            hit_node = inode;
-                            return hit_node;
-                        }
-                    }
-                }
-                break;
-            }
-            case SOURCE:
-            case SINK: {
-                float xcen, ycen;
-                draw_get_rr_src_sink_coords(rr_graph.rr_nodes()[inode], &xcen, &ycen);
-
-                // Now check if we clicked on this pin
-                if (click_x >= xcen - draw_coords->pin_size && click_x <= xcen + draw_coords->pin_size && click_y >= ycen - draw_coords->pin_size && click_y <= ycen + draw_coords->pin_size) {
-                    hit_node = inode;
-                    return hit_node;
-                }
-                break;
-            }
-            case CHANX:
-            case CHANY: {
-                bound_box = draw_get_rr_chan_bbox(inode);
-
-                // Check if we clicked on this wire, with 30%
-                // tolerance outside its boundary
-                const float tolerance = 0.3;
-                if (click_x >= bound_box.left() - tolerance && click_x <= bound_box.right() + tolerance && click_y >= bound_box.bottom() - tolerance && click_y <= bound_box.top() + tolerance) {
-                    hit_node = inode;
-                    return hit_node;
-                }
-                break;
-            }
-            default:
-                break;
-        }
-    }
-    return hit_node;
-}
-
-std::set<int> draw_expand_non_configurable_rr_nodes(int from_node) {
-    std::set<int> expanded_nodes;
-    draw_expand_non_configurable_rr_nodes_recurr(from_node, expanded_nodes);
-    return expanded_nodes;
-}
-
-void draw_expand_non_configurable_rr_nodes_recurr(int from_node,
-                                                  std::set<int>& expanded_nodes) {
-    auto& device_ctx = g_vpr_ctx.device();
-    const auto& rr_graph = device_ctx.rr_graph;
-    expanded_nodes.insert(from_node);
-
-    for (t_edge_size iedge = 0;
-         iedge < rr_graph.num_edges(RRNodeId(from_node)); ++iedge) {
-        bool edge_configurable = rr_graph.edge_is_configurable(RRNodeId(from_node), iedge);
-        int to_node = size_t(rr_graph.edge_sink_node(RRNodeId(from_node), iedge));
-
-        if (!edge_configurable && !expanded_nodes.count(to_node)) {
-            draw_expand_non_configurable_rr_nodes_recurr(to_node,
-                                                         expanded_nodes);
-        }
-    }
-}
-
-/* This routine is called when the routing resource graph is shown, and someone
- * clicks outside a block. That click might represent a click on a wire -- we call
- * this routine to determine which wire (if any) was clicked on.  If a wire was
- * clicked upon, we highlight it in Magenta, and its fanout in red.
- */
-static bool highlight_rr_nodes(float x, float y) {
-    t_draw_state* draw_state = get_draw_state_vars();
-
-    if (draw_state->draw_rr_toggle == DRAW_NO_RR && !draw_state->show_nets) {
-        application.update_message(draw_state->default_message);
-        application.refresh_drawing();
-        return false; //No rr shown
-    }
-
-    // Check which rr_node (if any) was clicked on.
-    int hit_node = draw_check_rr_node_hit(x, y);
-
-    return highlight_rr_nodes(hit_node);
-}
-
 #    if defined(X11) && !defined(__MINGW32__)
 void act_on_key_press(ezgl::application* /*app*/, GdkEventKey* /*event*/, char* key_name) {
     //VTR_LOG("Key press %c (%d)\n", key_pressed, keysym);
@@ -1811,165 +1563,9 @@ void act_on_mouse_move(ezgl::application* app, GdkEventButton* event, double x, 
     event = event; // just for hiding warning message
 }
 
-void draw_highlight_blocks_color(t_logical_block_type_ptr type,
-                                 ClusterBlockId blk_id) {
-    int k, iclass;
-    ClusterBlockId fanblk;
-
-    t_draw_state* draw_state = get_draw_state_vars();
-    auto& cluster_ctx = g_vpr_ctx.clustering();
-
-    for (k = 0; k < type->pb_type->num_pins; k++) { /* Each pin on a CLB */
-        ClusterNetId net_id = cluster_ctx.clb_nlist.block_net(blk_id, k);
-
-        if (net_id == ClusterNetId::INVALID())
-            continue;
-
-        auto physical_tile = physical_tile_type(blk_id);
-        int physical_pin = get_physical_pin(physical_tile, type, k);
-
-        iclass = physical_tile->pin_class[physical_pin];
-
-        if (physical_tile->class_inf[iclass].type == DRIVER) { /* Fanout */
-            if (draw_state->block_color(blk_id) == SELECTED_COLOR) {
-                /* If block already highlighted, de-highlight the fanout. (the deselect case)*/
-                draw_state->net_color[net_id] = ezgl::BLACK;
-                for (auto pin_id : cluster_ctx.clb_nlist.net_sinks(net_id)) {
-                    fanblk = cluster_ctx.clb_nlist.pin_block(pin_id);
-                    draw_reset_blk_color(fanblk);
-                }
-            } else {
-                /* Highlight the fanout */
-                draw_state->net_color[net_id] = DRIVES_IT_COLOR;
-                for (auto pin_id : cluster_ctx.clb_nlist.net_sinks(net_id)) {
-                    fanblk = cluster_ctx.clb_nlist.pin_block(pin_id);
-                    draw_state->set_block_color(fanblk, DRIVES_IT_COLOR);
-                }
-            }
-        } else { /* This net is fanin to the block. */
-            if (draw_state->block_color(blk_id) == SELECTED_COLOR) {
-                /* If block already highlighted, de-highlight the fanin. (the deselect case)*/
-                draw_state->net_color[net_id] = ezgl::BLACK;
-                fanblk = cluster_ctx.clb_nlist.net_driver_block(net_id); /* DRIVER to net */
-                draw_reset_blk_color(fanblk);
-            } else {
-                /* Highlight the fanin */
-                draw_state->net_color[net_id] = DRIVEN_BY_IT_COLOR;
-                fanblk = cluster_ctx.clb_nlist.net_driver_block(net_id); /* DRIVER to net */
-                draw_state->set_block_color(fanblk, DRIVEN_BY_IT_COLOR);
-            }
-        }
-    }
-
-    if (draw_state->block_color(blk_id) == SELECTED_COLOR) {
-        /* If block already highlighted, de-highlight the selected block. */
-        draw_reset_blk_color(blk_id);
-    } else {
-        /* Highlight the selected block. */
-        draw_state->set_block_color(blk_id, SELECTED_COLOR);
-    }
-}
-
-void deselect_all() {
-    // Sets the color of all clbs, nets and rr_nodes to the default.
-    // as well as clearing the highlighed sub-block
-
-    t_draw_state* draw_state = get_draw_state_vars();
-    auto& cluster_ctx = g_vpr_ctx.clustering();
-    auto& device_ctx = g_vpr_ctx.device();
-
-    /* Create some colour highlighting */
-    for (auto blk_id : cluster_ctx.clb_nlist.blocks()) {
-        if (blk_id != ClusterBlockId::INVALID())
-            draw_reset_blk_color(blk_id);
-    }
-
-    for (auto net_id : cluster_ctx.clb_nlist.nets())
-        draw_state->net_color[net_id] = ezgl::BLACK;
-
-    for (const RRNodeId& rr_id : device_ctx.rr_graph.nodes()) {
-        size_t i = (size_t)rr_id;
-        draw_state->draw_rr_node[i].color = DEFAULT_RR_NODE_COLOR;
-        draw_state->draw_rr_node[i].node_highlighted = false;
-    }
-    get_selected_sub_block_info().clear();
-}
-
 void draw_reset_blk_color(ClusterBlockId blk_id) {
     t_draw_state* draw_state = get_draw_state_vars();
     draw_state->reset_block_color(blk_id);
-}
-void draw_mux_with_size(ezgl::point2d origin, e_side orientation, float height, int size, ezgl::renderer* g) {
-    g->set_color(ezgl::YELLOW);
-    auto bounds = draw_mux(origin, orientation, height, g);
-
-    g->set_color(ezgl::BLACK);
-    g->draw_text(bounds.center(), std::to_string(size), bounds.width(),
-                 bounds.height());
-}
-
-//Draws a mux
-static inline ezgl::rectangle draw_mux(ezgl::point2d origin, e_side orientation, float height, ezgl::renderer* g) {
-    return draw_mux(origin, orientation, height, 0.4 * height, 0.6, g);
-}
-
-//Draws a mux, height/width define the bounding box, scale [0.,1.] controls the slope of the muxes sides
-static inline ezgl::rectangle draw_mux(ezgl::point2d origin, e_side orientation, float height, float width, float scale, ezgl::renderer* g) {
-    std::vector<ezgl::point2d> mux_polygon;
-
-    switch (orientation) {
-        case TOP:
-            //Clock-wise from bottom left
-            mux_polygon.push_back({origin.x - height / 2, origin.y - width / 2});
-            mux_polygon.push_back(
-                {origin.x - (scale * height) / 2, origin.y + width / 2});
-            mux_polygon.push_back(
-                {origin.x + (scale * height) / 2, origin.y + width / 2});
-            mux_polygon.push_back({origin.x + height / 2, origin.y - width / 2});
-            break;
-        case BOTTOM:
-            //Clock-wise from bottom left
-            mux_polygon.push_back(
-                {origin.x - (scale * height) / 2, origin.y - width / 2});
-            mux_polygon.push_back({origin.x - height / 2, origin.y + width / 2});
-            mux_polygon.push_back({origin.x + height / 2, origin.y + width / 2});
-            mux_polygon.push_back(
-                {origin.x + (scale * height) / 2, origin.y - width / 2});
-            break;
-        case LEFT:
-            //Clock-wise from bottom left
-            mux_polygon.push_back(
-                {origin.x - width / 2, origin.y - (scale * height) / 2});
-            mux_polygon.push_back(
-                {origin.x - width / 2, origin.y + (scale * height) / 2});
-            mux_polygon.push_back({origin.x + width / 2, origin.y + height / 2});
-            mux_polygon.push_back({origin.x + width / 2, origin.y - height / 2});
-            break;
-        case RIGHT:
-            //Clock-wise from bottom left
-            mux_polygon.push_back({origin.x - width / 2, origin.y - height / 2});
-            mux_polygon.push_back({origin.x - width / 2, origin.y + height / 2});
-            mux_polygon.push_back(
-                {origin.x + width / 2, origin.y + (scale * height) / 2});
-            mux_polygon.push_back(
-                {origin.x + width / 2, origin.y - (scale * height) / 2});
-            break;
-
-        default:
-            VTR_ASSERT_MSG(false, "Unrecognized orientation");
-    }
-    g->fill_poly(mux_polygon);
-
-    ezgl::point2d min((float)mux_polygon[0].x, (float)mux_polygon[0].y);
-    ezgl::point2d max((float)mux_polygon[0].x, (float)mux_polygon[0].y);
-    for (const auto& point : mux_polygon) {
-        min.x = std::min((float)min.x, (float)point.x);
-        min.y = std::min((float)min.y, (float)point.y);
-        max.x = std::max((float)max.x, (float)point.x);
-        max.y = std::max((float)max.y, (float)point.y);
-    }
-
-    return ezgl::rectangle(min, max);
 }
 
 ezgl::point2d tnode_draw_coord(tatum::NodeId node) {
