+++ vpr/src/draw/draw_toggle_functions.cpp
@@ -23,9 +23,10 @@
 #include "draw.h"
 #include "draw_rr.h"
 #include "draw_xtoy.h"
-#include "draw_triangle.h"
 #include "draw_toggle_functions.h"
-
+#include "draw_triangle.h"
+#include "draw_searchbar.h"
+#include "draw_mux.h"
 #include "read_xml_arch_file.h"
 #include "draw_global.h"
 #include "intra_logic_block.h"
